Refactor DTO packages
