+++ Backend/src/main/java/com/pg/mbti/dto/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.entity.login;
+package com.pg.mbti.dto;
 
 public record LoginResponseDto(String token) {
 }
