+++ Backend/src/main/java/com/pg/mbti/dto/UserProfileDto.java
@@ -1,4 +1,4 @@
-package com.pg.mbti.entity.profile;
+package com.pg.mbti.dto;
 
 public record UserProfileDto(
         String email,
